drop in rubyserial to replace serialport
This is because the serialport gem doesn't seem to work in later versions of ruby/operating systems, complaining about a missing symbol "rb_secure".  It's last update was in 2014 and uses a shared object file.  rubyserial doesn't require that compilation step and is a drop in replacement with the last update being in 2018.